Move reconciler/names.go to its own package
Avoids naming conflict between reconciler structs and reconciler
package.
